requirements.txt
audioread==3.0.1
av==12.2.0
Babel==2.15.0
bce-python-sdk==0.9.17
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
filelock==3.15.4
Flask==3.0.3
flask-babel==4.0.0
fonttools==4.53.1
fsspec==2024.6.1
future==1.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
macls==0.4.6
MarkupSafe==2.1.5
matplotlib==3.9.1
mpmath==1.3.0
msgpack==1.0.8
networkx==3.3
numba==0.60.0
numpy==1.26.4
packaging==24.1
paho-mqtt==2.1.0
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
protobuf==5.27.2
psutil==6.0.0
PyAudio==0.2.14
pycparser==2.22
pycryptodome==3.20.0
pydub==0.25.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
rarfile==4.2
requests==2.32.3
resampy==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
SoundCard==0.4.3
soundfile==0.12.1
soxr==0.3.7
sympy==1.12.1
termcolor==2.4.0
threadpoolctl==3.5.0
torch==2.3.1
torchaudio==2.3.1
torchinfo==1.8.0
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
visualdl==2.5.3
Werkzeug==3.0.3
